backend/pyproject.toml
[project]
name = "help-center-backend"
version = "0.1.0"
description = "Help Center Backend - Standalone documentation service"
requires-python = ">=3.12,<3.13"
license = { text = "AGPL-3.0" }
authors = [
    { name = "Pillar Team" }
]

dependencies = [
    # Core Django
    "django>=5.2.0,<5.3",
    "psycopg[binary]>=3.2.0",
    "python-dotenv>=1.0.0",
    # Django REST Framework
    "djangorestframework>=3.14.0",
    "djangorestframework-simplejwt>=5.3.0",
    "django-filter>=23.3",
    "django-cors-headers>=4.3.0",
    "drf-spectacular>=0.27.0",
    # Cache
    "redis>=5.0.1",
    "django-redis>=5.4.0",
    # Production server
    "gunicorn>=21.2.0",
    "whitenoise>=6.6.0",
    "uvicorn[standard]>=0.30.0",
    # Async support
    "asgiref>=3.7.0",
    "adrf>=0.1.6", # Async Django REST Framework views
    "channels>=4.0.0", # Django Channels for WebSocket support
    "channels-redis>=4.2.0", # Redis channel layer backend
    # LLM & AI
    "openai>=1.3.5",
    "litellm>=1.0.0",
    "cohere>=5.20.0",
    "google-generativeai>=0.8.5",
    # Document Processing & RAG
    "llama-index>=0.10.0",
    "llama-index-embeddings-openai>=0.1.0",
    "llama-index-embeddings-google-genai>=0.2.0",
    "pgvector>=0.2.4",
    "numpy>=1.24.0,<2.0",
    # Content Conversion
    "mistune>=3.0.0",
    "markdown>=3.5.0", # For rendering markdown to HTML
    # Web Scraping
    "firecrawl-py>=4.0.0",
    "requests>=2.31.0",
    "httpx>=0.26.0",
    "beautifulsoup4>=4.12.2",
    "html2text>=2020.1.16",
    # Phase 2: Content Sources (YouTube, Podcasts, Documents, Buckets)
    "youtube-transcript-api>=1.0.0", # YouTube transcript fetching
    "yt-dlp>=2024.0.0", # YouTube audio extraction
    "feedparser>=6.0.0", # RSS parsing for YouTube/Podcasts
    "pypdf>=4.0.0", # PDF text extraction
    "python-docx>=1.0.0", # DOCX text extraction
    "pydub>=0.25.0", # Audio preprocessing (needs ffmpeg)
    "groq>=0.9.0", # Whisper transcription
    "boto3>=1.35.0", # S3 client
    "google-cloud-storage>=2.18.0", # GCS client
    "django-storages[google]>=1.14.0", # Django storage backends for GCS/S3
    # Sources Integration
    "Unified-python-sdk>=0.48.0",
    "cryptography>=41.0.0",
    # Code Parsing (GitHub code indexing)
    "tree-sitter>=0.23.0",
    "tree-sitter-python>=0.23.0",
    "tree-sitter-javascript>=0.23.0",
    "tree-sitter-go>=0.23.0",
    "tree-sitter-typescript>=0.23.0",
    "tree-sitter-rust>=0.23.0",
    "tree-sitter-java>=0.23.0",
    "tree-sitter-ruby>=0.23.0",
    # Workflow Orchestration
    "hatchet-sdk>=1.20.0",
    # Development tools
    "ipython>=8.17.2",
    # watchdog is a transitive dependency (via uvicorn[standard])
    # Used by VSCode launch.json for Hatchet worker auto-reload
    # Testing
    "pytest>=7.4.3",
    "pytest-django>=4.7.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.23.0",
    "factory-boy>=3.3.0",
    "faker>=20.0.3",
    # Code quality
    "ruff>=0.1.6",
    "watchdog>=6.0.0",
    "tiktoken>=0.12.0",
    "ag-ui-protocol>=0.1.10",
    "pytest-xdist>=3.8.0",
    "uuid-utils>=0.14.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["apps", "common", "config"]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings.test"
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--verbose",
    "--tb=short",
    "--strict-markers",
    "-m", "not e2e",  # Skip e2e tests by default (require real API keys)
]
testpaths = ["tests"]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "e2e: End-to-end tests with real API calls (requires real API keys - use --ds=config.settings.e2e)",
    "workflow: marks tests as Hatchet workflow tests",
]
asyncio_mode = "auto"

[tool.ruff]
line-length = 100
target-version = "py312"
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by formatter)
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]
exclude = [
    ".git",
    ".ruff_cache",
    ".venv",
    "migrations",
    "__pycache__",
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]
